--- cracker-client/src/+components/Link.jsx
@@ -1,10 +1,0 @@
-import { Link as RouterLink } from "react-router-dom";
-import styled from "styled-components";
-
-export const Link = styled(RouterLink)`
-  color: ${(props) => props.theme.link};
-  text-decoration: none;
-  &:hover {
-    color: ${(props) => props.theme.linkHover};
-  }
-`;
