+++ cracker-client/src/+components/Navigation.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Link } from "./Link";
+import { Link } from "react-router-dom";
 import { withRouter } from "react-router-dom";
 import auth from "../+utils/Auth";
 import { Menu } from "antd";
