+++ cracker-client/src/users/+hooks/useUser.js
@@ -1,0 +1,13 @@
+import { gql } from "apollo-boost";
+import { useQuery } from "@apollo/react-hooks";
+
+const USERS = gql`
+  {
+    users {
+      firstName
+      lastName
+    }
+  }
+`;
+
+export const useUser = () => useQuery(USERS);
